+++ NEWS.md
@@ -1,3 +1,9 @@
+# gridmicrotex 0.0.2
+
+- Support the `\def` command
+- Bug fix `ggplot2` tile position.
+- `ggplot2` integration respects `latex_options`.
+
 # gridmicrotex 0.0.1
 
 Initial release.